# src/main/resources/api.yml
...
swagger: '2.0'
...
info:
  version: "0.0.1"
  title: Business Process API

# Describe your paths here
paths:
...
  /process-instance-groups/order-document:
    get:
      operationId: getOrderDocument
      parameters:
      - in: query
        name: processInstanceId
        type: string
        required: true
        description: Identifier of a process instance included in the group
      - name: Authorization
        in: header
        type: string
        required: true
      produces:
      - application/json
      responses:
        200:
          description: Retrieved the order document successfuly
        401:
          description: Invalid token. No user was found for the provided token
        404:
          description: No process instance or order exists
        500:
          description: Unexpected error while getting the order document
  /process-instance-groups/filters:
    get:
      operationId: getProcessInstanceGroupFilters
      parameters:
      - in: query
        name: partyId
        type: string
        description: Identifier of the party
      - in: query
        name: relatedProducts
        type: array
        description: Related products
      - in: query
        name: relatedProductCategories
        type: array
        description: Related product categories
      - in: query
        name: tradingPartnerIDs
        type: array
        description: Identifier of the corresponsing trading partner ID
      - in: query
        name: archived
        type: boolean
        default: false
      - in: query
        name: collaborationRole
        type: string
        required: false
      - in: query
        name: status
        required: false
        type: array
        items:
          type: string
          enum:
          - STARTED
          - CANCELLED
          - COMPLETED
      - name: Authorization
        in: header
        type: string
        required: true
      - in: query
        name: isProject
        type: boolean
        default: false
      produces:
      - application/json
      responses:
        200:
          description: Retrieved the ProcessInstanceGroupFilter successfully
          schema:
            $ref: '#/definitions/ProcessInstanceGroupFilter'
  /process-instance-groups/{id}:
    get:
      operationId: getProcessInstanceGroup
      parameters:
      - name: Authorization
        in: header
        type: string
        required: true
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: Retrieved the ProcessInstanceGroup successfully
          schema:
            $ref: '#/definitions/ProcessInstanceGroup'
        401:
          description: Invalid token. No user was found for the provided token
    delete:
      operationId: deleteProcessInstanceGroup
      parameters:
      - name: Authorization
        in: header
        type: string
        required: true
      - in: path
        name: id
        type: string
        required: true
      responses:
        200:
          description: Deleted the ProcessInstanceGroup successfully
        401:
          description: Invalid token. No user was found for the provided token
        404:
          description: There does not exist a process instance group with the given id
definitions:
...
  ProcessInstanceGroup:
    title: Process Instance Group Object
    type: object
    properties:
      ID:
        type: string
      name:
        type: string
      partyID:
        type: string
      dataChannelId:
        type: string
      processInstanceIDs:
        type: array
        items:
          type: string
      archived:
        type: boolean
      collaborationRole:
        type: string
      firstActivityTime:
        type: string
      lastActivityTime:
        type: string
      precedingProcessInstanceGroup:
        $ref: '#/definitions/ProcessInstanceGroup'
      status:
        type: string
        enum:
        - INPROGRESS
        - CANCELLED
...
  ProcessInstanceGroupFilter:
    title: Process Instance Group Filters
    description: Contains the filtering criteria for Process Instance Group objects
    type: object
    properties:
      startDate:
        type: string
      endDate:
        type: string
      tradingPartnerIDs:
        type: array
        items:
          type: string
      tradingPartnerNames:
        type: array
        items:
          type: string
      relatedProducts:
        type: array
        items:
          type: string
      relatedProductCategories:
        type: array
        items:
          type: string
      status:
        type: array
        items:
          type: string
          enum:
          - STARTED
          - CANCELLED
          - COMPLETED
...